separate grpc handler from grpc app
