+++ internal/grpchandler/file.go
@@ -1,4 +1,4 @@
-package grpcapp
+package grpchandler
 
 import (
 	"bytes"
@@ -9,59 +9,18 @@ import (
 
 	"github.com/go-seidon/hippo/api/grpcapp"
 	"github.com/go-seidon/hippo/internal/file"
-	"github.com/go-seidon/hippo/internal/healthcheck"
 	"github.com/go-seidon/provider/status"
-	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/metadata"
-	grpc_status "google.golang.org/grpc/status"
 )
-
-type healthHandler struct {
-	grpcapp.UnimplementedHealthServiceServer
-	healthClient healthcheck.HealthCheck
-}
-
-func (s *healthHandler) CheckHealth(ctx context.Context, p *grpcapp.CheckHealthParam) (*grpcapp.CheckHealthResult, error) {
-	checkRes, err := s.healthClient.Check(ctx)
-	if err != nil {
-		return nil, grpc_status.Error(codes.Unknown, err.Error())
-	}
-
-	details := map[string]*grpcapp.CheckHealthDetail{}
-	for _, item := range checkRes.Items {
-		details[item.Name] = &grpcapp.CheckHealthDetail{
-			Name:      item.Name,
-			Status:    item.Status,
-			CheckedAt: item.CheckedAt.UnixMilli(),
-			Error:     item.Error,
-		}
-	}
-
-	res := &grpcapp.CheckHealthResult{
-		Code:    checkRes.Success.Code,
-		Message: checkRes.Success.Message,
-		Data: &grpcapp.CheckHealthData{
-			Status:  checkRes.Status,
-			Details: details,
-		},
-	}
-	return res, nil
-}
-
-func NewHealthHandler(healthClient healthcheck.HealthCheck) *healthHandler {
-	return &healthHandler{
-		healthClient: healthClient,
-	}
-}
 
 type fileHandler struct {
 	grpcapp.UnimplementedFileServiceServer
-	fileService file.File
-	config      *GrpcAppConfig
+	fileClient file.File
+	config     *FileConfig
 }
 
 func (h *fileHandler) DeleteFileById(ctx context.Context, p *grpcapp.DeleteFileByIdParam) (*grpcapp.DeleteFileByIdResult, error) {
-	deletion, err := h.fileService.DeleteFile(ctx, file.DeleteFileParam{
+	deletion, err := h.fileClient.DeleteFile(ctx, file.DeleteFileParam{
 		FileId: p.FileId,
 	})
 	if err != nil {
@@ -83,7 +42,7 @@ func (h *fileHandler) DeleteFileById(ctx context.Context, p *grpcapp.DeleteFileB
 }
 
 func (h *fileHandler) RetrieveFileById(p *grpcapp.RetrieveFileByIdParam, stream grpcapp.FileService_RetrieveFileByIdServer) error {
-	retrieval, rerr := h.fileService.RetrieveFile(stream.Context(), file.RetrieveFileParam{
+	retrieval, rerr := h.fileClient.RetrieveFile(stream.Context(), file.RetrieveFileParam{
 		FileId: p.FileId,
 	})
 	if rerr != nil {
@@ -230,7 +189,7 @@ func (h *fileHandler) UploadFile(stream grpcapp.FileService_UploadFileServer) er
 		return nil
 	}
 
-	upload, uerr := h.fileService.UploadFile(
+	upload, uerr := h.fileClient.UploadFile(
 		stream.Context(),
 		file.WithFileInfo(
 			fileInfo.Name,
@@ -272,9 +231,18 @@ func (h *fileHandler) UploadFile(stream grpcapp.FileService_UploadFileServer) er
 	return nil
 }
 
-func NewFileHandler(fileService file.File, config *GrpcAppConfig) *fileHandler {
+type FileConfig struct {
+	UploadFormSize int64
+}
+
+type FileParam struct {
+	FileClient file.File
+	Config     *FileConfig
+}
+
+func NewFile(p FileParam) *fileHandler {
 	return &fileHandler{
-		fileService: fileService,
-		config:      config,
+		fileClient: p.FileClient,
+		config:     p.Config,
 	}
 }
