+++ internal/grpchandler/file_test.go
@@ -1,4 +1,4 @@
-package grpcapp_test
+package grpchandler_test
 
 import (
 	"context"
@@ -10,154 +10,19 @@ import (
 	mock_grpcapp "github.com/go-seidon/hippo/api/grpcapp/mock"
 	"github.com/go-seidon/hippo/internal/file"
 	mock_file "github.com/go-seidon/hippo/internal/file/mock"
-	"github.com/go-seidon/hippo/internal/grpcapp"
-	"github.com/go-seidon/hippo/internal/healthcheck"
-	mock_healthcheck "github.com/go-seidon/hippo/internal/healthcheck/mock"
+	"github.com/go-seidon/hippo/internal/grpchandler"
 	mock_context "github.com/go-seidon/provider/context/mock"
 	mock_io "github.com/go-seidon/provider/io/mock"
 	"github.com/go-seidon/provider/system"
 	"github.com/golang/mock/gomock"
-	"google.golang.org/grpc/codes"
 
 	"google.golang.org/grpc/metadata"
-	grpc_status "google.golang.org/grpc/status"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 )
 
 var _ = Describe("Handler Package", func() {
-
-	Context("CheckHealth function", Label("unit"), func() {
-		var (
-			handler       api.HealthServiceServer
-			healthService *mock_healthcheck.MockHealthCheck
-			ctx           context.Context
-			p             *api.CheckHealthParam
-			r             *api.CheckHealthResult
-			checkRes      *healthcheck.CheckResult
-		)
-
-		BeforeEach(func() {
-			t := GinkgoT()
-			ctrl := gomock.NewController(t)
-			healthService = mock_healthcheck.NewMockHealthCheck(ctrl)
-			handler = grpcapp.NewHealthHandler(healthService)
-			ctx = context.Background()
-			currentTs := time.Now().UTC()
-			p = &api.CheckHealthParam{}
-			r = &api.CheckHealthResult{
-				Code:    1000,
-				Message: "success check health",
-				Data: &api.CheckHealthData{
-					Status: "WARNING",
-					Details: map[string]*api.CheckHealthDetail{
-						"inet-conn": {
-							Name:      "inet-conn",
-							Status:    "OK",
-							Error:     "",
-							CheckedAt: currentTs.UnixMilli(),
-						},
-						"disk-check": {
-							Name:      "disk-check",
-							Status:    "FAILED",
-							Error:     "Critical: disk usage too high 61.93 percent",
-							CheckedAt: currentTs.UnixMilli(),
-						},
-					},
-				},
-			}
-			checkRes = &healthcheck.CheckResult{
-				Success: system.Success{
-					Code:    1000,
-					Message: "success check health",
-				},
-				Status: "WARNING",
-				Items: map[string]healthcheck.CheckResultItem{
-					"inet-conn": {
-						Name:      "inet-conn",
-						Status:    "OK",
-						Error:     "",
-						Fatal:     false,
-						CheckedAt: currentTs,
-					},
-					"disk-check": {
-						Name:      "disk-check",
-						Status:    "FAILED",
-						Error:     "Critical: disk usage too high 61.93 percent",
-						Fatal:     false,
-						CheckedAt: currentTs,
-					},
-				},
-			}
-		})
-
-		When("failed check service health", func() {
-			It("should return error", func() {
-				healthService.
-					EXPECT().
-					Check(gomock.Eq(ctx)).
-					Return(nil, &system.Error{
-						Code:    1001,
-						Message: "routine error",
-					}).
-					Times(1)
-
-				res, err := handler.CheckHealth(ctx, p)
-
-				Expect(res).To(BeNil())
-				Expect(err).To(Equal(
-					grpc_status.Error(codes.Unknown, "routine error"),
-				))
-			})
-		})
-
-		When("there is no states available", func() {
-			It("should return result", func() {
-				checkRes := &healthcheck.CheckResult{
-					Success: system.Success{
-						Code:    1000,
-						Message: "success check health",
-					},
-					Status: "OK",
-					Items:  map[string]healthcheck.CheckResultItem{},
-				}
-				healthService.
-					EXPECT().
-					Check(gomock.Eq(ctx)).
-					Return(checkRes, nil).
-					Times(1)
-
-				res, err := handler.CheckHealth(ctx, p)
-
-				r := &api.CheckHealthResult{
-					Code:    1000,
-					Message: "success check health",
-					Data: &api.CheckHealthData{
-						Status:  "OK",
-						Details: map[string]*api.CheckHealthDetail{},
-					},
-				}
-				Expect(res).To(Equal(r))
-				Expect(err).To(BeNil())
-			})
-		})
-
-		When("there are states available", func() {
-			It("should return result", func() {
-				healthService.
-					EXPECT().
-					Check(gomock.Eq(ctx)).
-					Return(checkRes, nil).
-					Times(1)
-
-				res, err := handler.CheckHealth(ctx, p)
-
-				Expect(res).To(Equal(r))
-				Expect(err).To(BeNil())
-			})
-		})
-	})
 
 	Context("DeleteFileById function", Label("unit"), func() {
 		var (
@@ -175,8 +40,10 @@ var _ = Describe("Handler Package", func() {
 			t := GinkgoT()
 			ctrl := gomock.NewController(t)
 			fileService = mock_file.NewMockFile(ctrl)
-			config := &grpcapp.GrpcAppConfig{}
-			handler = grpcapp.NewFileHandler(fileService, config)
+			handler = grpchandler.NewFile(grpchandler.FileParam{
+				FileClient: fileService,
+				Config:     &grpchandler.FileConfig{},
+			})
 			ctx = context.Background()
 			currentTs = time.Now()
 			p = &api.DeleteFileByIdParam{
@@ -303,8 +170,10 @@ var _ = Describe("Handler Package", func() {
 			t := GinkgoT()
 			ctrl := gomock.NewController(t)
 			fileService = mock_file.NewMockFile(ctrl)
-			config := &grpcapp.GrpcAppConfig{}
-			handler = grpcapp.NewFileHandler(fileService, config)
+			handler = grpchandler.NewFile(grpchandler.FileParam{
+				FileClient: fileService,
+				Config:     &grpchandler.FileConfig{},
+			})
 			ctx = mock_context.NewMockContext(ctrl)
 			p = &api.RetrieveFileByIdParam{
 				FileId: "file-id",
@@ -1029,10 +898,12 @@ var _ = Describe("Handler Package", func() {
 			t := GinkgoT()
 			ctrl := gomock.NewController(t)
 			fileService = mock_file.NewMockFile(ctrl)
-			config := &grpcapp.GrpcAppConfig{
-				UploadFormSize: 1073741824, //1GB
-			}
-			handler = grpcapp.NewFileHandler(fileService, config)
+			handler = grpchandler.NewFile(grpchandler.FileParam{
+				FileClient: fileService,
+				Config: &grpchandler.FileConfig{
+					UploadFormSize: 1073741824, //1GB
+				},
+			})
 			ctx = mock_context.NewMockContext(ctrl)
 			currentTs = time.Now()
 			stream = mock_grpcapp.NewMockFileService_UploadFileServer(ctrl)
@@ -1184,10 +1055,12 @@ var _ = Describe("Handler Package", func() {
 					Return(ctx).
 					Times(1)
 
-				config := &grpcapp.GrpcAppConfig{
-					UploadFormSize: 1,
-				}
-				handler := grpcapp.NewFileHandler(fileService, config)
+				handler := grpchandler.NewFile(grpchandler.FileParam{
+					FileClient: fileService,
+					Config: &grpchandler.FileConfig{
+						UploadFormSize: 1,
+					},
+				})
 
 				param := &api.UploadFileParam{
 					Data: &api.UploadFileParam_Chunks{
@@ -1230,10 +1103,12 @@ var _ = Describe("Handler Package", func() {
 					Return(ctx).
 					Times(1)
 
-				config := &grpcapp.GrpcAppConfig{
-					UploadFormSize: 1,
-				}
-				handler := grpcapp.NewFileHandler(fileService, config)
+				handler := grpchandler.NewFile(grpchandler.FileParam{
+					FileClient: fileService,
+					Config: &grpchandler.FileConfig{
+						UploadFormSize: 1,
+					},
+				})
 
 				param := &api.UploadFileParam{
 					Data: &api.UploadFileParam_Chunks{
